+++ examples/simple_example/main.tf
@@ -19,6 +19,10 @@ provider "google" {
   version     = "~> 1.20"
 }
 
+provider "local" {
+  version = "~> 1.2"
+}
+
 provider "null" {
   version = "~> 2.0"
 }
@@ -31,6 +35,45 @@ provider "random" {
   version = "~> 2.0"
 }
 
+resource "random_pet" "main" {
+  length    = "1"
+  prefix    = "forseti-simple-example"
+  separator = "-"
+}
+
+resource "google_compute_router" "main" {
+  name    = "${random_pet.main.id}"
+  network = "default"
+
+  bgp {
+    asn = "64514"
+  }
+
+  region  = "us-central1"
+  project = "${var.project_id}"
+}
+
+data "google_compute_subnetwork" "main" {
+  name    = "default"
+  project = "${var.project_id}"
+  region  = "${google_compute_router.main.region}"
+}
+
+resource "google_compute_router_nat" "main" {
+  name                               = "${random_pet.main.id}"
+  router                             = "${google_compute_router.main.name}"
+  nat_ip_allocate_option             = "AUTO_ONLY"
+  source_subnetwork_ip_ranges_to_nat = "LIST_OF_SUBNETWORKS"
+
+  subnetwork {
+    name                    = "${data.google_compute_subnetwork.main.self_link}"
+    source_ip_ranges_to_nat = ["ALL_IP_RANGES"]
+  }
+
+  project = "${var.project_id}"
+  region  = "${google_compute_router.main.region}"
+}
+
 module "forseti-install-simple" {
   source                   = "../../"
   project_id               = "${var.project_id}"
@@ -39,6 +82,14 @@ module "forseti-install-simple" {
   domain                   = "${var.domain}"
   client_instance_metadata = "${var.instance_metadata}"
   server_instance_metadata = "${var.instance_metadata}"
+  client_tags              = "${var.instance_tags}"
+  server_tags              = "${var.instance_tags}"
+  client_private           = "${var.private}"
+  server_private           = "${var.private}"
+  server_region            = "${google_compute_router_nat.main.region}"
+  client_region            = "${google_compute_router_nat.main.region}"
+  network                  = "${google_compute_router.main.network}"
+  subnetwork               = "${data.google_compute_subnetwork.main.name}"
 
   rules_path = "${path.module}/rules"