+++ test/fixtures/shared_vpc/main.tf
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
+provider "tls" {
+  version = "~> 1.2"
+}
+
 resource "tls_private_key" "main" {
   algorithm = "RSA"
   rsa_bits  = 4096
@@ -24,17 +28,26 @@ resource "local_file" "gce-keypair-pk" {
   filename = "${path.module}/sshkey"
 }
 
+module "bastion" {
+  source = "../bastion"
+
+  network    = "${var.network}"
+  project_id = "${var.network_project}"
+  subnetwork = "${var.subnetwork}"
+  zone       = "us-central1-f"
+}
+
 module "forseti-shared-vpc" {
-  source              = "../../../examples/shared_vpc"
-  credentials_path    = "${var.credentials_path}"
-  project_id          = "${var.project_id}"
-  region              = "${var.region}"
-  gsuite_admin_email  = "${var.gsuite_admin_email}"
-  network             = "${var.network}"
-  subnetwork          = "${var.subnetwork}"
-  network_project     = "${var.network_project}"
-  org_id              = "${var.org_id}"
-  domain              = "${var.domain}"
+  source             = "../../../examples/shared_vpc"
+  credentials_path   = "${var.credentials_path}"
+  project_id         = "${var.project_id}"
+  region             = "${var.region}"
+  gsuite_admin_email = "${var.gsuite_admin_email}"
+  network            = "${var.network}"
+  subnetwork         = "${var.subnetwork}"
+  network_project    = "${var.network_project}"
+  org_id             = "${var.org_id}"
+  domain             = "${var.domain}"
 
   instance_metadata {
     sshKeys = "ubuntu:${tls_private_key.main.public_key_openssh}"
@@ -50,9 +63,14 @@ resource "null_resource" "wait_for_server" {
     script = "${path.module}/scripts/wait-for-forseti.sh"
 
     connection {
-      user        = "ubuntu"
-      host        = "${module.forseti-shared-vpc.forseti-server-vm-public-ip}"
-      private_key = "${tls_private_key.main.private_key_pem}"
+      type                = "ssh"
+      user                = "ubuntu"
+      host                = "${module.forseti-shared-vpc.forseti-server-vm-ip}"
+      private_key         = "${tls_private_key.main.private_key_pem}"
+      bastion_host        = "${module.bastion.host}"
+      bastion_port        = "${module.bastion.port}"
+      bastion_private_key = "${module.bastion.private_key}"
+      bastion_user        = "${module.bastion.user}"
     }
   }
 }
@@ -66,9 +84,14 @@ resource "null_resource" "wait_for_client" {
     script = "${path.module}/scripts/wait-for-forseti.sh"
 
     connection {
-      user        = "ubuntu"
-      host        = "${module.forseti-shared-vpc.forseti-client-vm-public-ip}"
-      private_key = "${tls_private_key.main.private_key_pem}"
+      type                = "ssh"
+      user                = "ubuntu"
+      host                = "${module.forseti-shared-vpc.forseti-client-vm-ip}"
+      private_key         = "${tls_private_key.main.private_key_pem}"
+      bastion_host        = "${module.bastion.host}"
+      bastion_port        = "${module.bastion.port}"
+      bastion_private_key = "${module.bastion.private_key}"
+      bastion_user        = "${module.bastion.user}"
     }
   }
 }